PMM-6077 rename panel temporary return style workaround
