+++ dashboards/pmm-qan.json
@@ -147,7 +147,7 @@
                 }
             ],
             "title": "PMM Query Analytics",
-            "type": "pmm-qan"
+            "type": "pmm-qan-app-panel"
         }
     ],
     "refresh": false,
