+++ pmm-app/src/pmm-qan/module.ts
@@ -1,4 +1,4 @@
 import { PanelPlugin } from '@grafana/data';
-import QueryAnalyticsPanel from 'pmm-qan/panel/panel';
+import QueryAnalyticsPanel from 'pmm-qan/panel/qan';
 
 export const plugin = new PanelPlugin(QueryAnalyticsPanel);
