+++ pmm-app/src/pmm-qan/panel/components/Details/Details.service.ts
@@ -2,7 +2,7 @@ import {
   apiRequestManagement,
   apiRequestQAN,
 } from 'shared/components/helpers/api';
-import { getLabelQueryParams } from 'pmm-qan/panel/panel.tools';
+import { getLabelQueryParams } from 'pmm-qan/panel/qan.tools';
 
 
 const getExample = async ({
