+++ pmm-app/src/pmm-qan/panel/components/Filters/Filters.service.ts
@@ -1,5 +1,5 @@
 import { apiRequestQAN } from 'shared/components/helpers/api';
-import { getLabelQueryParams } from 'pmm-qan/panel/panel.tools';
+import { getLabelQueryParams } from 'pmm-qan/panel/qan.tools';
 
 const markCheckedLabels = (labels, paramLabels) => {
   Object.keys(labels).forEach((label) => {
