+++ pmm-app/src/pmm-qan/panel/components/Overview/Overview.service.ts
@@ -1,5 +1,5 @@
 import { apiRequestQAN } from 'shared/components/helpers/api';
-import { getLabelQueryParams } from 'pmm-qan/panel/panel.tools';
+import { getLabelQueryParams } from 'pmm-qan/panel/qan.tools';
 import { OverviewServiceInterface } from './Overview.types';
 
 export default {
