+++ pmm-app/src/pmm-qan/panel/components/Overview/Overview.tsx
@@ -7,8 +7,8 @@ import './Overview.scss';
 import { QueryAnalyticsProvider } from 'pmm-qan/panel/provider/provider';
 import 'shared/components/Elements/Spinner/Spinner';
 import { useOverviewTable } from './Overview.hooks';
-import { styles } from '../../panel.styles';
-import { DEFAULT_PAGE_NUMBER, DEFAULT_PAGE_SIZE, PAGE_SIZE_OPTIONS } from '../../panel.constants';
+import { styles } from '../../qan.styles';
+import { DEFAULT_PAGE_NUMBER, DEFAULT_PAGE_SIZE, PAGE_SIZE_OPTIONS } from '../../qan.constants';
 
 export const Overview: FC = () => {
   const [total, setTotal] = useState(30);
