+++ pmm-app/src/pmm-qan/panel/qan.constants.ts
