+++ pmm-app/src/pmm-qan/panel/qan.scss
