+++ pmm-app/src/pmm-qan/panel/qan.styles.ts
