+++ pmm-app/src/pmm-qan/panel/qan.tools.ts
