+++ pmm-app/src/shared/styles.scss
@@ -174,3 +174,11 @@ ol.breadcrumb-container {
 .app-tooltip:hover::before {
   display: inline-block;
 }
+
+.custom-grafana-plugin {
+  background-color: transparent !important;
+  height: auto;
+  min-height: 400px;
+  padding: 0px;
+  border: none;
+}
